Breaking nodes into their own files
